Add proper production chain for expressions
Expression produces AssignmentExpression which produces either
ConditionalExpression or Assignment. The LALR(1) Java grammar includes
in the Assignment production the non-terminal LeftHandSide which
produces Name, FieldAccess and ArrayAccess. This causes the generated
parser to reduce to Name, then to LeftHandSide, which makes the only
next acceptable token an assignment operator. Thus any name triggers the
parser to assume an assignment, which is bad.
If therefore purged the LeftHandSide nonterminal and instead included
Name directly in the Assignment production. I will therefore have to
include FieldAccess and ArrayAccess as alternative productions for
Assignment, too.